requirements.txt
streamlit
requests
llama-index
python-dotenv